Split the classification library into two pieces
